created a reusable button component that accepts a style prop to reduce code repetition
