+++ components/UI/ConfirmDeleteModal.js
@@ -1,9 +1,7 @@
 import { StyleSheet, Text, View, Modal, Pressable } from 'react-native';
-import React, { useState } from 'react';
+import React from 'react';
 import { ClockItColors } from '../../constants/styles';
-import ConfirmDeleteActivityButton from '../buttons/ConfirmDeleteActivityButton';
-import CancelDeleteButton from '../buttons/CancelDeleteButton';
-import MainUIButton from '../buttons/MainUIButton';
+import ReusableUIButton from '../buttons/ReusableUIButton';
 const ConfirmDeleteModal = ({ modalVisible, onCancelPress }) => {
   return (
     <Modal
@@ -20,8 +18,18 @@ const ConfirmDeleteModal = ({ modalVisible, onCancelPress }) => {
             </Text>
           </View>
           <View style={styles.functionContainer}>
-            <CancelDeleteButton onPress={onCancelPress} />
-            <ConfirmDeleteActivityButton onPress={() => {}} />
+            <ReusableUIButton
+              onPress={onCancelPress}
+              style={ReusableButtonStyles}
+              backgroundColor={ClockItColors.blue}>
+              Cancel
+            </ReusableUIButton>
+            <ReusableUIButton
+              onPress={() => {}}
+              backgroundColor={ClockItColors.confirmDelete}
+              style={ReusableButtonStyles}>
+              Delete activity
+            </ReusableUIButton>
           </View>
         </View>
       </Pressable>
@@ -41,14 +49,11 @@ const styles = StyleSheet.create({
   editingWarningAndFunctionContainer: {
     flex: 0.3,
     margin: 5,
-    // width: '80%',
     marginHorizontal: 24,
     backgroundColor: 'white',
     borderRadius: 10,
   },
-
   warningTextContainer: { flex: 1, padding: 20 },
-
   headerQuestion: {
     color: ClockItColors.darkestBlue,
     fontFamily: 'Manrope_700Bold',
@@ -62,15 +67,16 @@ const styles = StyleSheet.create({
     textAlign: 'center',
   },
   functionContainer: { flex: 0.5, flexDirection: 'row', justifyContent: 'center' },
-  modalContent: {
-    // backgroundColor: 'white',
-    // flex: 0.1,
-    // marginLeft: 22,
-    // marginBottom: 70,
-    // width: '50%',
-    // borderRadius: 8,
-    // shadowColor: '#090A49',
-    // shadowOffset: { width: 0, height: 4 },
-    // shadowRadius: 16,
+});
+
+const ReusableButtonStyles = StyleSheet.create({
+  button: { backgroundColor: ClockItColors.confirmDelete, flex: 0.5 },
+  buttonTextContainer: { flex: 1, justifyContent: 'center' },
+  buttonText: {
+    fontFamily: 'Manrope_500Medium',
+    color: 'white',
+    fontSize: 18,
+    textAlign: 'center',
+    letterSpacing: 1.1,
   },
 });
