--- components/buttons/MainUIButton.js
@@ -1,38 +1,0 @@
-import { Pressable, StyleSheet, Text, View } from 'react-native';
-
-function MainUIButton({ children, onPress, style }) {
-  return (
-    <Pressable
-      style={({ pressed }) => [styles.button, pressed && styles.pressed]}
-      onPress={onPress}>
-      <View>
-        <Text style={styles.buttonText}>{children}</Text>
-      </View>
-    </Pressable>
-  );
-}
-
-export default MainUIButton;
-
-const styles = StyleSheet.create({
-  button: {
-    borderRadius: 6,
-    paddingVertical: 6,
-    paddingHorizontal: 12,
-    backgroundColor: 'white',
-    elevation: 2,
-    shadowColor: 'black',
-    shadowOffset: { width: 1, height: 1 },
-    shadowOpacity: 0.25,
-    shadowRadius: 4,
-  },
-  pressed: {
-    opacity: 0.7,
-  },
-  buttonText: {
-    textAlign: 'center',
-    color: 'black',
-    fontSize: 16,
-    fontWeight: 'bold',
-  },
-});
